requirements.txt
Flask==2.0.1
pandas==1.3.3
requests==2.26.0
scikit-surprise==1.1.1
